+++ src/changelog.md
@@ -1,3 +1,11 @@
+#### 4.0.30
+* context menu items to copy values in clipboard, drilldown and drillthrough
+* ability to copy text value from the table cell
+* added ability to resize left table header
+* text wrap functionality added
+* tooltip for table when drilldown is available
+* button to call context menu (left top corner)
+
 #### 4.0.29
 * added autosearch after idle or input (#473)